Deprecate Self-Build Models (#10)
# Deprecate Self-Build Models

## :recycle: Current situation & Problem
At the start, we implemented all data models (following mhc and
engage-hf) on its own. This is not needed anymore as @PSchmiedmayer
pointed correctly out in #1.
all its FHIR wrappers (FHIRObservation, FHIRResource, etc.) were dead
code, and SchemaConverter was redundant with
@stanfordspezi/spezi-firebase-utils

## :gear: Release Notes
Consolidated types into functions/src/types/index.ts:
- Added User class (replacing the outdated simple interface that had
wrong field names like userType instead of type)
- Added UserAuth class + userAuthConverter
- Updated userConverter to use the correct full schema matching
Firestore documents

## :books: Documentation
Updated the readme


## :white_check_mark: Testing
N/A

## :pencil: Code of Conduct & Contributing Guidelines
By creating and submitting this pull request, you agree to follow our
[Code of
Conduct](https://github.com/StanfordSpezi/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordSpezi/.github/blob/main/CONTRIBUTING.md):
- [x] I agree to follow the [Code of
Conduct](https://github.com/StanfordSpezi/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordSpezi/.github/blob/main/CONTRIBUTING.md).
