--- functions/models/package-lock.json.license
@@ -1,4 +1,0 @@
-This source file is part of the Stanford Spezi Firebase Template project
-
-SPDX-FileCopyrightText: 2026 Stanford University and the project authors (see CONTRIBUTORS.md)
-SPDX-License-Identifier: MIT
