--- functions/models/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "@stanfordbdhg/spezi-firebase-models",
-  "version": "1.0.0",
-  "description": "Shared models for Spezi Firebase Template",
-  "main": "./lib/index.js",
-  "types": "./lib/index.d.ts",
-  "dependencies": {
-    "fhir": "^4.11.1",
-    "zod": "^4.0.5"
-  },
-  "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^6.12.0",
-    "@typescript-eslint/parser": "^6.12.0",
-    "eslint": "^8.54.0",
-    "typescript": "^5.2.2"
-  },
-  "scripts": {
-    "build": "tsc",
-    "build:watch": "tsc --watch",
-    "clean": "rm -rf lib",
-    "lint": "eslint --ext .js,.ts .",
-    "lint:fix": "eslint --ext .js,.ts . --fix"
-  },
-  "engines": {
-    "node": "18"
-  }
-}
