--- functions/models/src/codes/codes.ts
@@ -1,15 +1,0 @@
-// This source file is part of the Stanford Spezi Firebase Template project
-//
-// SPDX-FileCopyrightText: 2026 Stanford University and the project authors (see CONTRIBUTORS.md)
-// SPDX-License-Identifier: MIT
-
-export enum CodingSystem {
-  loinc = "http://loinc.org",
-  snomedCt = "http://snomed.info/sct",
-}
-
-export enum LoincCode {
-  stepCount = "41950-7",
-  bodyWeight = "29463-7",
-  heartRate = "8867-4",
-}
