--- functions/models/src/index.ts
@@ -1,21 +1,0 @@
-// This source file is part of the Stanford Spezi Firebase Template project
-//
-// SPDX-FileCopyrightText: 2026 Stanford University and the project authors (see CONTRIBUTORS.md)
-// SPDX-License-Identifier: MIT
-
-export * from "./codes/codes.js";
-export * from "./codes/quantityUnit.js";
-export * from "./fhir/fhirObservation.js";
-export * from "./fhir/fhirQuestionnaireResponse.js";
-export * from "./fhir/fhirConverters.js";
-export * from "./helpers/schemaConverter.js";
-export * from "./types/observationQuantity.js";
-export * from "./types/user.js";
-export * from "./types/userAuth.js";
-export * from "./types/userType.js";
-export * from "./types/userMessage.js";
-
-export type UserObservationCollection =
-  | "stepCount"
-  | "bodyWeight"
-  | "heartRate";
