--- functions/models/src/types/observationQuantity.ts
@@ -1,12 +1,0 @@
-// This source file is part of the Stanford Spezi Firebase Template project
-//
-// SPDX-FileCopyrightText: 2026 Stanford University and the project authors (see CONTRIBUTORS.md)
-// SPDX-License-Identifier: MIT
-
-import { type QuantityUnit } from "../codes/quantityUnit.js";
-
-export interface ObservationQuantity {
-  date: Date;
-  value: number;
-  unit: QuantityUnit;
-}
