--- functions/models/src/types/userType.ts
@@ -1,10 +1,0 @@
-// This source file is part of the Stanford Spezi Firebase Template project
-//
-// SPDX-FileCopyrightText: 2026 Stanford University and the project authors (see CONTRIBUTORS.md)
-// SPDX-License-Identifier: MIT
-
-export enum UserType {
-  owner = "owner",
-  clinician = "clinician",
-  patient = "patient",
-}
