--- functions/models/tsconfig.json
@@ -1,23 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "ESNext",
-    "noImplicitReturns": true,
-    "noUnusedLocals": true,
-    "outDir": "lib",
-    "sourceMap": true,
-    "strict": true,
-    "target": "ES2020",
-    "moduleResolution": "node",
-    "allowSyntheticDefaultImports": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "resolveJsonModule": true,
-    "declaration": true,
-    "declarationMap": true
-  },
-  "compileOnSave": true,
-  "include": [
-    "src"
-  ]
-}
