+++ functions/src/functions/updateUserInformation.ts
@@ -3,13 +3,13 @@
 // SPDX-FileCopyrightText: 2026 Stanford University and the project authors (see CONTRIBUTORS.md)
 // SPDX-License-Identifier: MIT
 
-import { UserType } from "@stanfordbdhg/spezi-firebase-models";
 import { getFirestore } from "firebase-admin/firestore";
 import { z } from "zod/v4";
 import { validatedOnCall } from "../helpers/validatedOnCall.js";
 import { Credential } from "../services/auth/credential.js";
 import { DefaultDatabaseService } from "../services/database/databaseService.js";
 import { DefaultUserService } from "../services/user/defaultUserService.js";
+import { UserType } from "../types/index.js";
 
 const updateUserInformationInputSchema = z.object({
   userId: z.string(),
