+++ functions/src/services/user/userService.ts
@@ -3,7 +3,7 @@
 // SPDX-FileCopyrightText: 2026 Stanford University and the project authors (see CONTRIBUTORS.md)
 // SPDX-License-Identifier: MIT
 
-import { type UserAuth, type User } from "@stanfordbdhg/spezi-firebase-models";
+import { type UserAuth, type User } from "../../types/index.js";
 import { type Document } from "../database/databaseService.js";
 
 export interface UserService {
